Authentication/ player stats
Player can now log on and authentication requests are sent through and gateway server and then an authentication server and tells the player if they can load into the game server.
Upon logging in, players and their stats are loaded into the game server and sent to the player to update their stats.